am: change the type of _uptime() from uintptr_t to unsigned long

See merge request !116

+++ am/arch/mips32-minimal/include/arch.h
@@ -22,7 +22,7 @@ extern "C" {
 #endif
 
 static inline void _ioe_init() { }
-static inline uintptr_t _uptime() { return 0; }
+static inline unsigned long _uptime() { return 0; }
 static inline void _draw_rect(const uint32_t *pixels, int x, int y, int w, int h) {}
 static inline void _draw_sync() {}
